Move ModelParams into main namespace
