+++ fwdpy11/__init__.py
@@ -25,4 +25,4 @@ if sys.version_info[0] < 3:
 from fwdpy11._version import __version__ # NOQA
 from ._fwdpy11 import * # NOQA
 from ._dev import *
-
+from ._model_params import *
