+++ tests/test_DiploidPopulation.py
@@ -88,7 +88,7 @@ class testPythonObjects(unittest.TestCase):
         self.rng = fwdpy11.GSLrng(101)
 
     def testParentalData(self):
-        from fwdpy11.model_params import ModelParams
+        from fwdpy11 import ModelParams
         from fwdpy11.wright_fisher import evolve
         params = ModelParams(**self.pdict)
         evolve(self.rng, self.pop, params)
@@ -107,7 +107,7 @@ class testPythonObjects(unittest.TestCase):
             self.assertEqual(i.nodes[1], fwdpy11.NULL_NODE)
 
     def test_nodes_after_evolution(self):
-        from fwdpy11.model_params import ModelParams
+        from fwdpy11 import ModelParams
         from fwdpy11.wright_fisher import evolve
         params = ModelParams(**self.pdict)
         evolve(self.rng, self.pop, params)
@@ -116,7 +116,7 @@ class testPythonObjects(unittest.TestCase):
             self.assertEqual(i.nodes[1], fwdpy11.NULL_NODE)
 
     def testMutationLookupTable(self):
-        from fwdpy11.model_params import ModelParams
+        from fwdpy11 import ModelParams
         from fwdpy11.wright_fisher import evolve
         params = ModelParams(**self.pdict)
         evolve(self.rng, self.pop, params)
@@ -127,7 +127,7 @@ class testPythonObjects(unittest.TestCase):
                 self.assertTrue(i in lookup[self.pop.mutations[i].pos])
 
     def testMutationIndices(self):
-        from fwdpy11.model_params import ModelParams
+        from fwdpy11 import ModelParams
         from fwdpy11.wright_fisher import evolve
         params = ModelParams(**self.pdict)
         evolve(self.rng, self.pop, params)
